refactored to use controllers (more modular)
